+++ src/app/common/form-input-field.tsx
@@ -2,10 +2,10 @@ import { FieldError, UseFormRegister } from "react-hook-form";
 
 export type FormInputFieldProps = {
   name: ValidFieldNames;
+  onBlur?: any;
   placeholder: string;
   register: UseFormRegister<any>;
   error: FieldError | undefined;
-  valueAsNumber?: boolean;
 };
 
 export type ValidFieldNames =
@@ -21,17 +21,19 @@ export type ValidFieldNames =
 
 const FormInputField: React.FC<FormInputFieldProps> = ({
   name,
+  onBlur,
   placeholder,
   register,
   error,
-  valueAsNumber,
 }) => (
   <>
     <input
       className="ml-1 mb-3 border-2 border-slate-300"
       type="text"
       placeholder={placeholder}
-      {...register(name, { valueAsNumber })}
+      {...register(name, {
+        onBlur,
+      })}
     />
     {error && <span className="error-message">{error.message}</span>}
   </>